+++ Continuum Emulator/Emulator/Execution/ExDEBUG.cs
@@ -16,16 +16,8 @@ namespace Continuum93.Emulator.Execution
                 ClientActions.StartStepByStepMode();
             }
 
-            //tempLog += $"State:{Machine.COMPUTER.CPU.FREGS.GetRegistersState()}{Environment.NewLine}";
+            // Enter Service mode here
 
-
-
-            // Dump regs
-            //Debug.WriteLine(Machine.COMPUTER.CPU.REGS.EFG);
-            //string path = Machine.COMPUTER.MEMC.GetStringAt(Machine.COMPUTER.CPU.REGS.ABCD);
-            //byte[] bytes1 = Machine.COMPUTER.MEMC.DumpMemAt(Machine.COMPUTER.CPU.REGS.CDE, 8);
-            //byte[] bytes2 = Machine.COMPUTER.MEMC.DumpMemAt(Machine.COMPUTER.CPU.REGS.FGH, 8);
-            //Debug.WriteLine(Machine.COMPUTER.CPU.REGS.RST);
         }
     }
 }