+++ Continuum Emulator/CHANGELOG.md
@@ -10,6 +10,11 @@
 - Small refactor on the AND instructions. Nothing essential changed, just cleaned up the codebase a bit.
 - In preparation to deprecating Continuum Tools: Implemented an integrated Service Mode overlay (accessible always with F1) with a smooth animated transition
 - The emulator now renders using an aspect-correct destination rectangle with automatic letterboxing/pillarboxing to preserve the 16:9 ratio on any resolution, and it automatically switches between pixel-perfect PointClamp scaling for integer multiples and LinearClamp for non-integer scales to avoid visual artifacts on awkward resolutions.
+- Started working on a new integrated "Service Mode" to eventually replace Continuum Tools. Currently, it supports:
+  - Toggling to service with F1;
+  - Disassembler view;
+  - Configurable windows layout;
+  - work in progress...
 
 ## 2.1.212 — 25.10.2025